Add test+fix for bug that caused inner joins to be used even when in the context of 'or' scopes where the property wasn't actually required.
Technically we could architect this in a way that allowed a slightly higher percentage of INNER JOINs to be used (i.e. if the chain of associations from the parent domain to the current association are all non-nullable, then it doesn't matter that they were used in an 'or' scope--they could still be joined using an INNER JOIN.) However, due to the the complexity of that (and especially the need for considering how that might interact with sub-query embedded scopes) for right now, I'll push that off into the future in favor of fixing this bug more immediately.